doesMatchPattern moved to its own util as match-glob
