+++ ui/app/abilities/variable.js
@@ -2,6 +2,7 @@
 import { computed, get } from '@ember/object';
 import { or } from '@ember/object/computed';
 import AbstractAbility from './abstract';
+import { default as doesMatchPattern } from 'nomad-ui/utils/match-glob';
 
 const WILDCARD_GLOB = '*';
 const WILDCARD_PATTERN = '/';
@@ -188,7 +189,7 @@ export default class Variable extends AbstractAbility {
     const matches = [];
 
     for (const pathName of pathNames) {
-      if (this._doesMatchPattern(pathName, path)) matches.push(pathName);
+      if (doesMatchPattern(pathName, path)) matches.push(pathName);
     }
 
     return matches;
@@ -207,41 +208,41 @@ export default class Variable extends AbstractAbility {
     return this._smallestDifference(allMatchingPaths, path);
   }
 
-  _doesMatchPattern(pattern, path) {
-    const parts = pattern?.split(WILDCARD_GLOB);
-    const hasLeadingGlob = pattern?.startsWith(WILDCARD_GLOB);
-    const hasTrailingGlob = pattern?.endsWith(WILDCARD_GLOB);
-    const lastPartOfPattern = parts[parts.length - 1];
-    const isPatternWithoutGlob = parts.length === 1 && !hasLeadingGlob;
+  // _doesMatchPattern(pattern, path) {
+  //   const parts = pattern?.split(WILDCARD_GLOB);
+  //   const hasLeadingGlob = pattern?.startsWith(WILDCARD_GLOB);
+  //   const hasTrailingGlob = pattern?.endsWith(WILDCARD_GLOB);
+  //   const lastPartOfPattern = parts[parts.length - 1];
+  //   const isPatternWithoutGlob = parts.length === 1 && !hasLeadingGlob;
 
-    if (!pattern || !path || isPatternWithoutGlob) {
-      return pattern === path;
-    }
+  //   if (!pattern || !path || isPatternWithoutGlob) {
+  //     return pattern === path;
+  //   }
 
-    if (pattern === WILDCARD_GLOB) {
-      return true;
-    }
+  //   if (pattern === WILDCARD_GLOB) {
+  //     return true;
+  //   }
 
-    let subPathToMatchOn = path;
-    for (let i = 0; i < parts.length; i++) {
-      const part = parts[i];
-      const idx = subPathToMatchOn?.indexOf(part);
-      const doesPathIncludeSubPattern = idx > -1;
-      const doesMatchOnFirstSubpattern = idx === 0;
+  //   let subPathToMatchOn = path;
+  //   for (let i = 0; i < parts.length; i++) {
+  //     const part = parts[i];
+  //     const idx = subPathToMatchOn?.indexOf(part);
+  //     const doesPathIncludeSubPattern = idx > -1;
+  //     const doesMatchOnFirstSubpattern = idx === 0;
 
-      if (i === 0 && !hasLeadingGlob && !doesMatchOnFirstSubpattern) {
-        return false;
-      }
+  //     if (i === 0 && !hasLeadingGlob && !doesMatchOnFirstSubpattern) {
+  //       return false;
+  //     }
 
-      if (!doesPathIncludeSubPattern) {
-        return false;
-      }
+  //     if (!doesPathIncludeSubPattern) {
+  //       return false;
+  //     }
 
-      subPathToMatchOn = subPathToMatchOn.slice(0, idx + path.length);
-    }
+  //     subPathToMatchOn = subPathToMatchOn.slice(0, idx + path.length);
+  //   }
 
-    return hasTrailingGlob || path.endsWith(lastPartOfPattern);
-  }
+  //   return hasTrailingGlob || path.endsWith(lastPartOfPattern);
+  // }
 
   _computeLengthDiff(pattern, path) {
     const countGlobsInPattern = pattern
